==> docs/01_DataLoad.rmd <==
---
title: "01 Data Load"
author: J Andres Gannon, Center for Peace and Security Studies (cPASS)
date: "`r Sys.Date()`"
output:
  html_document:
    theme: flatly
    code_download: yes
    code_folding: hide
    toc: yes
    number_sections: yes
    toc_float:
      collapsed: yes
  html_notebook:
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
editor_options:
  chunk_output_type: inline
---
<style>
    body .main-container {
        max-width: 100%;
    }
</style>

This is the entry point for the paper "After Deterrence: Explaining Conflict Short of War". This document creates the raw data as a .rds file pulled from the googlesheet where the data structure was manually created.

# Load packages
Pipe operators don't play nice and must be manually loaded up front. All other package functions are extracted using ::
```{r}
library("magrittr")
```

# Load data
We load the data from the google spreadsheet tab that has aggregated the info from previous versions. Each row is now a gray zone case and the columns are coded covariates. The distance data represented here was calculated using the code in 02_Distance.Rmd and then imputed into the google spreadsheet.

For replication, the googlesheet is located in data/raw so access permissions are not a problem
```{r}
grayzone <- read.csv(paste0(here::here(), "/data/raw/RussianWars-new_cases.csv"))
```
# Save file
New .rds file is saved so the raw data remains unmodified by future changes
```{r}
saveRDS(grayzone, paste0(here::here(), "/data/","grayzone_cpass.rds"))
```
