==> docs/03a_PriorData_ICB.Rmd <==
---
title: "03a PriorData: ICB"
author: J Andres Gannon, Center for Peace and Security Studies (cPASS)
date: "`r Sys.Date()`"
output:
  html_document:
    theme: flatly
    code_download: yes
    code_folding: hide
    toc: yes
    number_sections: yes
    toc_float:
      collapsed: yes
  html_notebook:
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
editor_options:
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This document will look at coverage of Russian military interventions from 1991-2015 in the ICB data and identify variables of interest. It is the first of 3 prior dataset used to identify the universe of cases.

The ICB data comes from Brecher, Michael and Jonathan Wilkenfeld (2000). A Study of Crisis. Ann Arbor: University of Michigan Press.

# Preparation
## Load packages
Pipe operators have trouble loading for individual commands
```{r}
library(magrittr)
```

## Load data
The ICB data exists in two forms. ICB1 is system-level data and ICB2 is actor-level data. We'll primarily be using the actor-level data but will load both. Data files not created by the authors are stored in inst/extdata/
```{r}
# Load system-level data
icb_sys <- read.csv(file = paste0(here::here(), '/inst/extdata/icb1v12.csv'))

# Load actor-level data
icb_actor <- read.csv(file = paste0(here::here(), '/inst/extdata/icb2v12.csv'))
```

## Subset
We are only interests in cases since 1991 where Russia was an actor (primarily as an initiator)
```{r}
# Subset to post-1991 cases
icb_actor_rus <- subset(icb_actor, yrtrig > "1990")

# Subset to cases involving Russia
icb_actor_rus <- subset(icb_actor_rus, actor == "RUS")

crisis_count <- nrow(icb_actor_rus)
crisis_list <- icb_actor_rus$crisname
```

There are `r crisis_count` ICB crises post-1991 in which Russia was a crisis actor. This includes `r crisis_list`.

# Add variables
## Target state
Add a column for the name of the target state since target state-year will be used to merge all the datasets
```{r}
icb_actor_rus$crisname

icb_actor_rus$target <- NA

icb_actor_rus$target[icb_actor_rus$crisname == 'PANKISI GORGE'] <- 'Georgia'
icb_actor_rus$target[icb_actor_rus$crisname == 'S. OSSETIA/ABKHAZIA'] <- 'Georgia'
icb_actor_rus$target[icb_actor_rus$crisname == 'RUSSO-GEORGIAN WAR'] <- 'Georgia'
icb_actor_rus$target[icb_actor_rus$crisname == 'CRIMEA-DONBASS'] <- 'Ukraine'
icb_actor_rus$target[icb_actor_rus$crisname == 'TURKEY-RUSSIA JET INCIDENT'] <- 'Syria'
```

# Summary statistics
## Choose variables
Identify variables of interest regarding the means employed, intensity of crises, casualties, etc. From looking over the codebook that includes the following:

* CENVIO = centrality of violence
  1-4, least to most

* SEVVIO = intensity of violence
  1-4, least to most

* OUTCOM = content of crisis outcome
  1-4, victory to defeat

* OUTESR = escalation of reduction of tension
  1 escalation, 2 de-escalation

* TRGTERRA = elapsed time between perception of trigger and termination (count in days)

* GEOG = geographic location of crisis
  - (9) Central Asia
  - (10) West Asia (Turkey, Armenia, Azerbaijan, Georgia and Cyprus)
  - (11) East Asia
  - (12) South-East Asia
  - (13) South Asia
  - (15) Middle East (countries from the UN’s West Asia category not listed above)
  - (20) West Africa
  - (21) North Africa
  - (22) East Africa
  - (23) Southern Africa
  - (24) Central Africa
  - (30) Euro-Asia (Russia)
  - (31) East Europe
  - (32) Central Europe (Germany, East Germany, Austria, Switzerland)
  - (33) West Europe
  - (34) North Europe
  - (35) South Europe
  - (41) North America
  - (42) Central America (including Caribbean countries)
  - (43) South America
  - (51) Australasia (including Oceania countries)

* CRACTLOC = distance of crisis actor from location of crisis
  1-4, closest to farthest

* NOACTR = number of actors involved (count)

* VIOL = violence associated with crisis actor
  1-4, least to most
```{r}
# Keep identifier variables
vars_id <- c("crisno", "systrgyr", "crisname", "target")
vars_interest <- c("cenvio", "sevvio", "outcom", "outesr", "trgterra", "geog", "cractloc", "noactr", "viol")

vars_keep <- c(vars_id, vars_interest)

icb_actor_rus <- icb_actor_rus[ , vars_keep]

# Reorder the variables
icb_actor_rus <- icb_actor_rus[ , c('crisno', 'crisname', 'systrgyr', 'target', 'cenvio', 'sevvio', 'viol', 'outesr', 'outcom', 'trgterra', 'geog', 'cractloc', 'noactr')]

# Rename variables for readability
icb_actor_rus <- icb_actor_rus %>% dplyr::rename('Crisis No' = crisno,
                                                 'Year' = systrgyr,
                                                 'Crisis Name' = crisname,
                                                 'Target' = target,
                                                 'Crisis Violence Centrality' = cenvio,
                                                 'Crisis Violence Severity' = sevvio,
                                                 'Outcome' = outcom,
                                                 'Escalation' = outesr,
                                                 'Duration (days)' = trgterra,
                                                 'Location' = geog,
                                                 'Distance from Actor' = cractloc,
                                                 'No. of Actors' = noactr,
                                                 'Actor Violence Severity' = viol)

# Recode some variables to be more intuitive
icb_actor_rus$Escalation <- car::recode(icb_actor_rus$Escalation, "2 = 0; 3 = NA")
icb_actor_rus$Location <- car::recode(icb_actor_rus$Location, "10 = 'West Asia'; 31 = 'East Europe'")
icb_actor_rus$Outcome <- car::recode(icb_actor_rus$Outcome, "1 = 'Victory'; 2 = 'Compromise'; 3 = 'Stalemate'; 4 = 'Defeat'")
icb_actor_rus$`Distance from Actor` <- car::recode(icb_actor_rus$`Distance from Actor`, "1 = 'Home territory'; 2 = 'Sub-region'; 3 = 'Same continent'")
```

## Summary
Make a table summarizing the data on Russian crisis participation
```{r}
formattable::formattable(icb_actor_rus,
                         align = c("l", "l", "c", "c", "c", "c", "c", "c", "c", "c", "l", "l", "c"),
                         list(formattable::area(col = c('Crisis Violence Centrality', 'Crisis Violence Severity', 'Actor Violence Severity')) 
                              ~ formattable::color_tile("red", "red4")))
```

# Save final dataframe
```{r}
saveRDS(icb_actor_rus, paste0(here::here(), "/data/","grayzone_icb.rds"))
```
