==> docs/05a_Model_Prep.Rmd <==
---
title: "05a Model Preparation"
author: J Andres Gannon
date: "`r Sys.Date()`"
output:
  html_document:
    theme: flatly
    code_download: yes
    code_folding: hide
    toc: yes
    number_sections: yes
    toc_float:
      collapsed: yes
  html_notebook:
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
editor_options:
  chunk_output_type: inline
---

```{r setup, include=FALSE}
library(magrittr)
library(ggplot2)
options(scipen = 999)
knitr::opts_chunk$set(warning = FALSE, message = FALSE, echo = TRUE)
```

# Load
## New data
Load the newly created data of Russia interventions post-1994.
```{r}
df <- readRDS(paste0(here::here(), '/data/grayzone_aggregate_new.rds'))
```

## European states
Load a blank dataset of European countries from 1994-2018. We use the list of COW states. The time series is constant for all countries except for Montenegro which exists from 2006 onwards and Kosovo which exists from 2008 onward.
```{r}
full <- read.csv(paste0(here::here(), "/inst/extdata/system2016.csv")) %>%
  dplyr::select(-version)

# Subset to Europe and code Georgia, Kosovo, Turkey, Yugoslavia, and Czech Republic as Europe
full$continent <- countrycode::countrycode(full$ccode, 'cown', 'continent')
full$continent[full$stateabb == "GRG"] <- "Europe"
full$continent[full$stateabb == "KOS"] <- "Europe"
full$continent[full$stateabb == "TUR"] <- "Europe"
full$continent[full$stateabb == "YUG"] <- "Europe"
full$continent[full$stateabb == "CZE" | full$stateabb == "CZR"] <- "Europe"

full <- full %>%
  dplyr::filter(continent == "Europe" | stateabb == "CAN" | stateabb == "SAU" | stateabb == "SYR" | stateabb == "USA") %>% # Include non-European states in the original data
  dplyr::filter(year > 1993) %>%
  dplyr::select(-continent)

# Add 2017 and 2018 for all states
countries <- as.character(sort(unique(full$ccode)))
newyr <- data.frame(expand.grid(ccode = c(countries),
                 year = 2017:2018))
newyr$ccode <- as.numeric(as.character(newyr$ccode))
newyr$stateabb <- countrycode::countrycode(newyr$ccode, "cown", "cowc")

full <- rbind(full, newyr)

full <- full %>%
  dplyr::rename(country_code_cow = ccode)

DT::datatable(plyr::count(full, "stateabb"))
DT::datatable(plyr::count(full, "year"))
```

## Subset of relevant European states
We want to include all European countries that a) have conflict history with Russia between 1945-1993, b) were former Soviet states, c) are contiguous with Russia. We add a fourth criteria of states that are targets of Russian attacks at any point during our time period, which is unique from the other 3 for Austria, Malta, Spain, and Germany. We include separate measures for inclusion in both criteria for robustness
```{r}
# Conflict history
## MID
mid <- haven::read_dta(paste0(here::here(),"/inst/extdata/MID/MIDB 4.3.dta"))

# Subset to columns of interest
mid <- mid %>%
  dplyr::select(dispnum3, ccode, styear) %>%
  dplyr::filter(styear > 1945 & styear < 1994) %>%
  dplyr::select(-styear) %>%
  dplyr::group_by(dispnum3) %>%
  dplyr::filter(any(ccode == 365)) %>%
  dplyr::ungroup() %>%
  dplyr::filter(!ccode == 365) %>%
  dplyr::select(-dispnum3) %>%
  dplyr::distinct() %>%
  dplyr::mutate(disphist = 1) %>%
  dplyr::rename(country_code_cow = ccode)

## ICB
icb <- read.csv(file = paste0(here::here(), '/inst/extdata/icb2v12.csv'))

icb <- icb %>%
  dplyr::select(crisno, cracid, yrtrig) %>%
  dplyr::filter(yrtrig > 1945 & yrtrig < 1994) %>%
  dplyr::select(-yrtrig) %>%
  dplyr::group_by(crisno) %>%
  dplyr::filter(any(cracid == 365)) %>%
  dplyr::ungroup() %>%
  dplyr::filter(!cracid == 365) %>%
  dplyr::select(-crisno) %>%
  dplyr::distinct() %>%
  dplyr::rename(country_code_cow = cracid) %>%
  dplyr::mutate(disphist = 1)

## Merge the two
disphist <- dplyr::bind_rows(mid, icb) %>%
  dplyr::distinct()

# FSU
fsu <- read.csv(paste0(here::here(), "/data/membership.csv"))

fsu <- fsu %>%
  dplyr::select(state, fsu, wp) %>%
  dplyr::filter(fsu == 1 | wp == 1) %>%
  dplyr::mutate(fsu_wp = fsu == 1 | wp == 1)
fsu$fsu_wp[fsu$fsu_wp == "TRUE"] <- 1

fsu$country_code_cow <- countrycode::countrycode(fsu$state, "country.name", "cown")  

fsu <- fsu %>%
  dplyr::select(country_code_cow, fsu_wp)

# Contiguous
contig <- haven::read_dta(paste0(here::here(),"/inst/extdata/DirectContiguity320/contdird.dta"))

contig <- contig %>%
  dplyr::filter(state1no == 365) %>%
  dplyr::mutate(contig = conttype < 4) %>%
  dplyr::select(state2no, year, contig) %>%
  dplyr::rename(country_code_cow = state2no) %>%
  dplyr::filter(contig == "TRUE")

contig$year <- as.integer(contig$year)

# In DCID, REI, or new intervention data
cyber <- df

cyber$country_code_cow <- countrycode::countrycode(cyber$target, "country.name", "cown")

cyber <- cyber %>%
  dplyr::mutate(interv = resp_convmil == 1 | 
                  resp_convmil_gro == 1 | 
                  resp_convmil_airsea == 1 | 
                  resp_paramil == 1 | 
                  resp_cyberdisrup == 1 | 
                  resp_infoops == 1) %>%
  dplyr::select(country_code_cow, interv) %>%
  dplyr::distinct()
cyber$interv[cyber$interv == "TRUE"] <- 1

cyber$continent <- countrycode::countrycode(cyber$country_code_cow, 'cown', 'continent')
cyber$continent[cyber$country_code_cow == "372"] <- "Europe" # Georgia
cyber$continent[cyber$country_code_cow == "347"] <- "Europe" # Kosovo
cyber$continent[cyber$country_code_cow == "640"] <- "Europe" # Turkey
cyber$continent[cyber$country_code_cow == "345"] <- "Europe" # Yugoslavia
cyber$continent[cyber$country_code_cow == "315" | cyber$stateabb == "316"] <- "Europe" # Czech Republic

# Combine them all with the main data
full <- dplyr::left_join(full, disphist)
full <- dplyr::left_join(full, fsu)
full <- dplyr::left_join(full, contig)
full <- dplyr::left_join(full, cyber)

# Fix continents again, since they got lost in the merge
full$continent <- countrycode::countrycode(full$country_code_cow, 'cown', 'continent')
full$continent[full$country_code_cow %in% c("315", # Czech Republic
                                            "316", # Czech Republic
                                            "345", # Yugoslavia
                                            "347", # Kosovo
                                            "372", # Georgia
                                            "640" #Turkey
                                            )] <- "Europe"


full <- full %>%
  dplyr::mutate(relevant_conserv = (continent == "Europe" & (disphist == 1 | fsu_wp == 1 | contig == 1))) %>%
  dplyr::mutate(relevant_expansive = (disphist == 1 | fsu_wp == 1 | contig == 1 | interv == 1)) %>%
  dplyr::select(stateabb, country_code_cow, year, relevant_conserv, relevant_expansive)

full$relevant_conserv[full$relevant_conserv == "TRUE"] <- "1"
full$relevant_conserv[is.na(full$relevant_conserv)] <- "0"
full$relevant_conserv <- as.numeric(full$relevant_conserv)

full$relevant_expansive[full$relevant_expansive == "TRUE"] <- "1"
full$relevant_expansive[is.na(full$relevant_expansive)] <- "0"
full$relevant_expansive <- as.numeric(full$relevant_expansive)
```

# Clean
Drop columns that describe the coding process. This table shows a list of the 60 unique target-years in the data. It drops 22 cases that are repeat Russian interventions with the same target-year.
```{r}
# Subset and replace NAs with 0s for intensity variables
df <- df %>% dplyr::select(target, year_start, dplyr::starts_with("resp_")) %>%
  dplyr::rename(year = year_start) %>%
  dplyr::mutate(target = dplyr::recode(target, US = "United States", UK = "United Kingdom")) %>%
  tidyr::replace_na(list(resp_convmil = 0, resp_convmil_gro = 0, resp_convmil_airsea = 0, resp_paramil = 0, resp_cyberdisrup = 0, resp_infoops = 0))
df$target <- stringr::str_trim(df$target)

# Give ordinal value to each level
df$resp_convmil_gro <- df$resp_convmil_gro * 5
df$resp_convmil_airsea <- df$resp_convmil_airsea * 4
df$resp_paramil <- df$resp_paramil * 3
df$resp_cyberdisrup <- df$resp_cyberdisrup * 2
df$resp_infoops <- df$resp_infoops * 1

# Keep highest value per year
for( i in 1:nrow(df)){
  df$intensity[i] <- max(df$resp_convmil_gro[i],
                         df$resp_convmil_airsea[i],
                         df$resp_paramil[i], 
                         df$resp_cyberdisrup[i], 
                         df$resp_infoops[i])
}

# Subset to the highest value for country-year
df <- df %>%
  dplyr::group_by(target, year) %>%
  dplyr::filter(intensity == max(intensity)) %>%
  dplyr::select(target, year, intensity) %>%
  dplyr::distinct()
```

Merge the cleaned gray zone data into the Europe country-year dataset. This drops the Chechnya 1999 observation since they are not a recognized European state in the COW dataset
```{r}
# Add cow codes to gray zone data for merging
df$country_code_cow <- countrycode::countrycode(df$target, "country.name", "cown")
df$country_code_cow <- as.numeric(df$country_code_cow)
df <- df %>%
  dplyr::ungroup() %>%
  dplyr::select(country_code_cow, year, intensity)

# Merge
df <- dplyr::full_join(full, df)

# Fix state names and abbrevations that were lost in the merging
df$stateabb <- countrycode::countrycode(df$country_code_cow, "cown", "cowc")
df$statename <- countrycode::countrycode(df$country_code_cow, "cown", "country.name")

# Set non-events as 0
df$intensity[is.na(df$intensity)] <- 0

# Set countries that don't meet our 3-prong criteria as 0. Here, that primarily means non-European targets of Russian attacks (US, Canada, Saudi Arabia, etc) as well as some European states like Malta, Spain, Germany, and Austria. Do the same for the expansive criteria
df$relevant_conserv[is.na(df$relevant_conserv)] <- 0
df$relevant_expansive[is.na(df$relevant_expansive)] <- 0

# Drop non-countries, which is just Chechnya
df <- df %>%
  dplyr::filter(!is.na(country_code_cow))

# Drop Russia since we are not counting attacks on self
df <- df %>%
  dplyr::filter(!country_code_cow == 365)
```

# Rename vars
```{r}
df <- df %>%
  dplyr::rename(cname1 = statename, cabbrev1 = stateabb, ccode1 = country_code_cow) %>%
  dplyr::select(year, cname1, cabbrev1, ccode1, intensity, relevant_conserv, relevant_expansive)

df$ccode1 <- as.numeric(df$ccode1)
```

# Add variables
We add a battery of variables for the RHS of the model. SideB is always Russia, so we can hard-code that variable.
```{r}
# COW codes
df$cname2 <- "Russia"
df$cabbrev2 <- "RUS"
df$ccode2 <- 365

# Re-order the vars to make things more aesthetically pleasing
df <- df %>%
  dplyr::select(year, cname1, cabbrev1, ccode1, cname2, cabbrev2, ccode2, relevant_conserv, relevant_expansive, intensity)
```

## Distance
### From US
This is an independent variable of interest, coding how far the target state is from the US as the geographic proxy for deterrence. We code capital to capital distance using the 2022 release of the cshapes package.
```{r}
# US based distance
df_dist <- read.table(paste0(here::here(), '/inst/extdata/cshapes_2.0_dist_COW.csv'), header = TRUE, sep = ",") %>%
  dplyr::filter(ccode2 == 2 &
                  year >= 1994) %>%
  dplyr::select(year, ccode1, dplyr::ends_with("dist")) %>%
  dplyr::rename(mindist_us = mindist,
                capdist_us = capdist,
                centdist_us = centdist)

df <- dplyr::left_join(df, df_dist)

# Fix edge cases
## Kosovo KOS 347. We take the 2090 distance and apply it to 2008
df$mindist_us[df$cabbrev1 == "KOS"] <- 6571.54
df$capdist_us[df$cabbrev1 == "KOS"] <- 7790.441
df$centdist_us[df$cabbrev1 == "KOS"] <- 9197.661

## Montenegro MNG 341
df$mindist_us[df$cabbrev1 == "MNG"] <- 6450.735
df$capdist_us[df$cabbrev1 == "MNG"] <- 7675.717
df$centdist_us[df$cabbrev1 == "MNG"] <- 9108.594
```

### From Russia
We code distance from Russia to control for Russian resolve using the same data as above.
```{r}
# Russia based distance
df_dist <- read.table(paste0(here::here(), '/inst/extdata/cshapes_2.0_dist_COW.csv'), header = TRUE, sep = ",") %>%
  dplyr::filter(ccode2 == 365 &
                  year >= 1994) %>%
  dplyr::select(year, ccode1, dplyr::ends_with("dist")) %>%
  dplyr::rename(mindist_rus = mindist,
                capdist_rus = capdist,
                centdist_rus = centdist)

df <- dplyr::left_join(df, df_dist)

# Edge cases
## Kosovo KOS 347. We take the 2090 distance and apply it to 2008
df$mindist_rus[df$cabbrev1 == "KOS"] <- 1220.7768
df$capdist_rus[df$cabbrev1 == "KOS"] <- 1875.417
df$centdist_rus[df$cabbrev1 == "KOS"] <- 5247.687

## Montenegro MNG 341
df$mindist_rus[df$cabbrev1 == "MNG"] <- 1212.709
df$capdist_rus[df$cabbrev1 == "MNG"] <- 1984.534
df$centdist_rus[df$cabbrev1 == "MNG"] <- 5312.238
```

### From NATO headquarters (Brussels) contingent on distance from Russia
NATO HQ is in Brussels (capital of Belgium). and Supreme Headquarters Allied Powers Europe (SHAPE) and Allied Command Operations (ACO) are also in Belgium at Casteau and Mons.

The NATO Defense College: Navigating Between Critical Analysis, Strategic Education, and Partnerships
Trine Villumsen Berling and Brooke A. Smith-Windsor
p 163-186

```{r}
# NATO based capital distance
df_dist <- read.table(paste0(here::here(), '/inst/extdata/cshapes_2.0_dist_COW.csv'), header = TRUE, sep = ",") %>%
  dplyr::filter(ccode2 == 211 &
                  year >= 1994) %>%
  dplyr::select(year, ccode1, capdist) %>%
  dplyr::rename(capdist_natohq = capdist)

df <- dplyr::left_join(df, df_dist)

# Edge cases
## Kosovo KOS 347. We take the 2090 distance and apply it to 2008
df$capdist_natohq[df$cabbrev1 == "KOS"] <- 1568.424

## Montenegro MNG 341
df$capdist_natohq[df$cabbrev1 == "MNG"] <- 1470.748

# Create new var of relative NATO HQ to Russia distance. Higher values mean a state is closer to NATO than to Moscow
df$capdist_natorelative <- df$capdist_rus - df$capdist_natohq
```

### From any NATO member
Calculate each state's distance from its nearest NATO member. NATO states are coded as distance 0
```{r}
df_dist <- read.table(paste0(here::here(), '/inst/extdata/cshapes_2.0_dist_COW.csv'), header = TRUE, sep = ",") %>%
  dplyr::filter(year >= 1994) %>%
  dplyr::select(year, ccode1, ccode2, capdist)

nato <- read.csv(paste0(here::here(), "/data/membership.csv")) %>%
  dplyr::mutate(ccode2 = countrycode::countrycode(state, "country.name", "cown")) %>%
  dplyr::select(ccode2, join_nato_year)

df_dist <- dplyr::left_join(df_dist, nato) %>%
  dplyr::mutate(ccode2_nato = ifelse(join_nato_year <= year, 1, 0)) %>%
  dplyr::filter(ccode2_nato == 1)

df_natodist <- df_dist %>%
  dplyr::select(year, ccode1)

df_natodist <- dplyr::left_join(df_natodist, df_dist) %>%
  dplyr::group_by(year, ccode1) %>%
  dplyr::slice(which.min(capdist)) %>%
  dplyr::mutate(capdist_natomin = capdist) %>%
  dplyr::select(year, ccode1, capdist_natomin)

df <- dplyr::left_join(df, df_natodist)
```

## NATO membership
We code NATO membership as a binary variable for each country-year
```{r}
# Load NATO membership data
nato <- read.csv(paste0(here::here(), "/data/membership.csv")) %>%
  dplyr::select(state, dplyr::starts_with("join_"))

# Create dataframe of all of these countries from 1948 (year before NATO was founded) to 2018 (last year in our data)
nato_long <- nato %>%
  dplyr::filter(!is.na(join_nato_year)) %>%
  dplyr::select(state) %>%
  dplyr::mutate(start_year = 1948,
                end_year = 2018) %>%
  dplyr::mutate(year = purrr::map2(start_year, end_year, `:`)) %>% 
  dplyr::select(-start_year, -end_year) %>%
  tidyr::unnest() %>%
  dplyr::select(year, state)

nato_long$nato_stage <- NA
nato_long$nato_stage <- as.character(nato_long$nato_stage)

nato_long <- dplyr::left_join(nato_long, nato) %>%
  dplyr::mutate(nato_stage = dplyr::if_else(year == join_pfp_year, "PfP", nato_stage),
                nato_stage = dplyr::if_else(year == join_intdial_year & is.na(nato_stage), "IntDial", nato_stage),
                nato_stage = dplyr::if_else(year == join_map_year & is.na(nato_stage), "MAP", nato_stage),
                nato_stage = dplyr::if_else(year == join_nato_year & is.na(nato_stage), "Full", nato_stage)) %>%
  dplyr::mutate(nato_stage = dplyr::if_else(join_nato_year < year & is.na(nato_stage), "Full", nato_stage),
                nato_stage = dplyr::if_else(is.na(join_pfp_year) & 
                                              is.na(join_intdial_year) &
                                              is.na(join_map_year) &
                                              year < join_nato_year,
                                            "Non-member", nato_stage),
                nato_stage = dplyr::if_else(is.na(nato_stage) &
                                              year < join_pfp_year,
                                            "Non-member", nato_stage)) %>%
  tidyr::fill(nato_stage) %>%
  dplyr::select(year, state, nato_stage)

# Create COW code column for merging
nato_long$ccode1 <- countrycode::countrycode(nato_long$state, "country.name", "cown")
nato_long <- nato_long %>%
  dplyr::select(-state)

nato$ccode1 <- countrycode::countrycode(nato$state, "country.name", "cown")
nato <- nato %>%
  dplyr::select(ccode1, join_nato_year)

# Merge both
df <- dplyr::left_join(df, nato_long)
df <- dplyr::left_join(df, nato)

# Clean
df <- df %>%
  dplyr::mutate(nato_stage = dplyr::if_else(is.na(nato_stage), "Non-member", nato_stage)) %>%
  dplyr::mutate(nato_memb_duration = year - join_nato_year) %>%
  dplyr::mutate(nato_memb = ifelse(nato_memb_duration >= 0, 1, 0)) %>%
  dplyr::mutate(nato_memb_duration = ifelse(nato_memb_duration < 0, 0, nato_memb_duration)) %>%
  dplyr::mutate(nato_prememb = dplyr::if_else(nato_stage %in% c("PfP", "IntDial", "MAP") , 1, 0)) %>%
  dplyr::mutate(nato_prememb = tidyr::replace_na(nato_prememb, 0)) %>%
  dplyr::rename(NATOmem_MEM = nato_memb, 
                NATOdur_MEM = nato_memb_duration, 
                NATOjoined_MEM = join_nato_year) %>%
  dplyr::select(-NATOjoined_MEM)

df$NATOdur_MEM[is.na(df$NATOdur_MEM)] <- 0
df$NATOmem_MEM[is.na(df$NATOmem_MEM)] <- 0
```

## DCAs
We use Kinne's measure of defense cooperation agreements which has more variation than NATO membership
```{r, eval = FALSE}
dcid <- read.csv(paste0(here::here(), "/inst/extdata/kinne/DCAD-v1.0-dyadic.csv")) %>%
  dplyr::filter(year > 1993,
                ccode2 == 2) %>%
  dplyr::select(year, ccode1, dcaGeneralV1) %>%
  dplyr::rename(dca = dcaGeneralV1)

df <- dplyr::left_join(df, dcid)
```

## Alliance depth
We use Benson and Clinton's measure of alliance depth to get more fine-grained distinctions among alliance commitments
```{r, eval = FALSE}
# depth <- read.csv(paste0(here::here(), "/inst/extdata/gannon_kent/AllianceDataScore_Kent.csv"))
depth <- read.csv(paste0(here::here(), "/inst/extdata/gannon_kent/AllianceDataScore.csv"))

## Use the year, country-code and depth for merging. All rows concern alliance depth with the United States
depth <- depth %>%
  dplyr::select(V1, depth.score) %>%
  dplyr::rename(atopid = V1,
                alliance_depth = depth.score)

# Get country code IDs based off ATOP IDs
atop <- read.csv(paste0(here::here(), "/inst/extdata/ATOP 5_0/atop5_0dy.csv"))

atop <- atop %>%
  dplyr::select(atopid1, year, mem1, mem2) %>%
  dplyr::rename(atopid = atopid1)

depth_dyad <- dplyr::left_join(atop, depth)

depth_dyad <- depth_dyad %>%
  dplyr::filter(mem1 == 2) %>%
  dplyr::filter(year > 1993) %>%
  dplyr::select(-mem1, -atopid, -year) %>%
  dplyr::distinct() %>%
  dplyr::rename(ccode1 = mem2) %>%
  dplyr::mutate(ccode1 = as.numeric(ccode1))

# Merge with original data
df <- dplyr::left_join(df, depth_dyad)
```

## Troop basing
There is data on US troop basing provided by Allen, Michael A., Michael E. Flynn, and Carla Martinez Machain. 2021. “Global U.S. military deployment data: 1950-2020.” Working Paper
```{r, eval = FALSE}
# Pull US troop deployment data at the country level and merge 
troops <- troopdata::get_troopdata(startyear = 1994,
                         endyear = 2018) %>%
  dplyr::select(year, ccode, troops) %>%
  dplyr::rename(ccode1 = ccode,
                ustroops = troops)

df <- dplyr::left_join(df, troops)

# Pull US base data at the base level, aggregate to country level, and merge
bases <- troopdata::get_basedata(country_count = TRUE) %>%
  dplyr::select(ccode, base, lilypad) %>%
  dplyr::group_by(ccode) %>%
  dplyr::summarise(usbases_main = sum(base),
                   usbases_lilypads = sum(lilypad),
                   usbases_all = sum(base) + sum(lilypad)) %>%
  dplyr::rename(ccode1 = ccode)

df <- dplyr::left_join(df, bases)
df$usbases_all[is.na(df$usbases_all)] <- 0
df$usbases_main[is.na(df$usbases_main)] <- 0
df$usbases_lilypads[is.na(df$usbases_lilypads)] <- 0
```

## CINC scores
We code CINC scores to control for power disparity between the two actors. Constitent with conventional measures, we also code the ratio of the challenger's (Russia) material capabilities ot the sum of both states capabilities
```{r}
cinc <- read.csv(paste0(here::here(), '/inst/extdata/NMC_5_0.csv'))

cinc_ccode1 <- cinc %>% dplyr::select(ccode, year, cinc) %>%
  dplyr::rename(ccode1 = ccode, cinc1 = cinc)

df <- dplyr::left_join(df, cinc_ccode1)

cinc_ccode2 <- cinc %>% dplyr::select(ccode, year, cinc) %>%
  dplyr::rename(ccode2 = ccode, cinc2 = cinc)

df <- dplyr::left_join(df, cinc_ccode2)

df$cinc_ratio <- df$cinc1/(df$cinc1+df$cinc2)
```

## Military expenditure
Since CINC data only goes up to 2012, we supplement it with SIPRI military expenditure data which covers the full temporal span of the data
```{r}
# Mil exp data
sipri <- read.csv(paste0(here::here(), "/inst/extdata/sipri_milex.csv"))
df <- dplyr::left_join(df, sipri)

# Mil exp per capita
sipri <- read.csv(paste0(here::here(), "/inst/extdata/sipri_milexpercap.csv"))
df <- dplyr::left_join(df, sipri)
```

## Regime type
We code polity scores to control for whether or not both sides are democracies. In alignment with Gartzke and Jo 2009, they are re-scales from -10-10 to 0-20 so allow for a multiplicative variable. We also convert each to a dummy variable for democracy if polity is greater than 6.

There is a discrepancy over how Polity uses COW coes for Kosovo and the successor states to Yugoslavia. We defer to the COW codings and alter the polity numeric codings to match the polity name abbreviations that describe each observation
```{r}
polity <- readxl::read_xls(paste0(here::here(), "/inst/extdata/p5v2018.xls"))

polity$polity2_scaled <- polity$polity2 + 10

polity_ccode1 <- polity %>% dplyr::select(ccode, year, polity2_scaled, polity2) %>%
  dplyr::mutate(demo1 = ifelse(polity2 >= 6, 1, 0)) %>%
  dplyr::rename(ccode1 = ccode, polity1 = polity2_scaled) %>%
  dplyr::select(-polity2)

# Fix edge cases
## Polity miscodes the COW code for Serbia and Montenegro as 347 when it should remain 345 for Yugoslavia (347 refers to Kosovo)
polity_ccode1$ccode1[polity_ccode1$ccode1 == 347] <- 345

## Polity miscodes the COW code for Kosovo as 341 when it should be 347 (341 refers to Montenegro)
polity_ccode1$ccode1[polity_ccode1$ccode1 == 341] <- 347

## Polity miscodes the COW code for Serbia as 342 when it should be 345 (342 does not exist in COW)
polity_ccode1$ccode1[polity_ccode1$ccode1 == 342] <- 345

## Polity miscodes the COW code for Montenegro as 348 when it should be 341 (348 does not exist in COW)
polity_ccode1$ccode1[polity_ccode1$ccode1 == 348] <- 341


df <- dplyr::left_join(df, polity_ccode1)

polity_ccode2 <- polity %>% dplyr::select(ccode, year, polity2_scaled, polity2) %>%
  dplyr::mutate(demo2 = ifelse(polity2 > 6, 1, 0)) %>%
  dplyr::select(-polity2) %>%
  dplyr::rename(ccode2 = ccode, polity2 = polity2_scaled)
df <- dplyr::left_join(df, polity_ccode2)

df$polity_intaxn <- df$polity1 * df$polity2
```

## Civil war
We use UCDP to code whether a country-year is involved in an intrastate war
```{r}
load(paste0(here::here(), "/inst/extdata/ucdp-prio-acd-201.RData"))
ucdp <- ucdp_prio_acd_201
ucdp <- ucdp %>% dplyr::select(gwno_a, year, type_of_conflict) %>%
  dplyr::rename(ccode1 = gwno_a) %>%
  dplyr::mutate(ccode1 = strsplit(as.character(ccode1), ",")) %>%
  tidyr::unnest(ccode1) %>%
  dplyr::filter(type_of_conflict == 3) %>%
  dplyr::mutate(civilwar = 1) %>%
  dplyr::select(-type_of_conflict)

ucdp$ccode1 <- as.numeric(stringr::str_trim(ucdp$ccode1))
ucdp <- ucdp %>% dplyr::distinct()
df <- dplyr::left_join(df, ucdp)
df$civilwar[is.na(df$civilwar)] <- 0
```

## Nuclear status
We code nuclear status of the target state using the coding from from Jo and Gartzke 2007
```{r}
df$nuclear1 <- 0
df$nuclear1[df$cname1 == "France" | 
              df$cname1 == "United Kingdom" |
              df$cname1 == "United States"] <- 1
```

## GDP and population
We code GDP and GDP per capita of Russia and the target state using the World Bank's World Development Indicator (WDI).
```{r}
ipe <- WDI::WDI(indicator = c('NY.GDP.PCAP.KD', 'NY.GDP.MKTP.KD', 'SP.POP.TOTL'),
                country = "all",
                start = 1994,
                end = 2018)

# Target state
ipe$ccode1 <- countrycode::countrycode(ipe$iso2c, "iso2c", "cown")

ipe_ccode1 <- ipe %>%
  dplyr::rename(gdppc1_2010const = "NY.GDP.PCAP.KD", gdp1_2010const = "NY.GDP.MKTP.KD", pop1 = "SP.POP.TOTL") %>%
  dplyr::select(year, ccode1, gdppc1_2010const, gdp1_2010const, pop1)

# World Bank has Yugoslavia coded as Serbia and Kosovo has no COW code, so we fix those before merging
ipe_ccode1$ccode1[ipe$country == "Serbia"] <- 345
ipe_ccode1$ccode1[ipe$country == "Kosovo"] <- 347

df <- dplyr::left_join(df, ipe_ccode1)

# Non-target state
ipe$ccode2 <- countrycode::countrycode(ipe$iso2c, "iso2c", "cown")

ipe_ccode2 <- ipe %>%
  dplyr::rename(gdppc2_2010const = "NY.GDP.PCAP.KD", gdp2_2010const = "NY.GDP.MKTP.KD", pop2 = "SP.POP.TOTL") %>%
  dplyr::select(year, ccode2, gdppc2_2010const, gdp2_2010const, pop2)

df <- dplyr::left_join(df, ipe_ccode2)
```

## Country age
We calculate how long a country has existed, since newness may proxy for state instability or other conditions that make intervention by a great power more likely. We use COW data for the years a country existed. Although that only goes as far back as 1816, it should be sufficient given the effect of this variable is not linear
```{r}
# Load COW data of country-years in the international system
age <- read.csv(paste0(here::here(), "/inst/extdata/system2016.csv")) %>%
  dplyr::select(-version)

# Find earliest year for each country
age <- age %>%
  dplyr::group_by(ccode) %>%
  dplyr::top_n(-1) %>%
  dplyr::select(-stateabb) %>%
  dplyr::rename(startyr = year, ccode1 = ccode)

# Merge
df <- dplyr::left_join(df, age)

# Create new var for country age
df <- df %>%
  dplyr::mutate(age1 = (year - startyr)) %>%
  dplyr::select(-startyr)
```

## ELF
We take a measure of ethno-linguistic fractionalization (ELF) from the Historical Index of Ethnic Fractionalization (HEIF) dataset that covers 1945-2013. The index measures the probability that 2 randomly drawn individuals are not from the same ethnic group.

HEIF is missing 2 states prior to 2014 that are included in Fearon and Laitin's measure of ethnic and linguistic fractionalization; France and Yugoslavia. We fill in those missing values for 1993-2014. 
```{r}
# Load HEIF data
elf <- read.csv(paste0(here::here(), "/inst/extdata/HIEF_data.csv"))

# Load Fearon and Laitin data
fl <- DirectEffects::civilwar %>%
  dplyr::select(ccode, cname, year, ef) %>%
  dplyr::filter(cname == "FRANCE" | cname == "YUGOSLAV") %>%
  dplyr::filter(year >= 1993)

# Match COW codes
elf$ccode1 <- countrycode::countrycode(elf$Country, "country.name", "cown")

# Fix edge cases
elf$ccode1[elf$Country == "Serbia"] <- 345

# Rename vars
elf <- elf %>%
  dplyr::rename(year = Year, elf = EFindex) %>%
  dplyr::select(-Country)

# Merge
df <- dplyr::left_join(df, elf)

# Impute from Fearon and Laitin
df$elf[df$year > 2014 & df$ccode1 == 220] <- 0.2720740 # France
df$elf[df$year > 2014 & df$ccode1 == 345] <- 0.5747164 # Yugoslavia

```

# Fix edge cases
Fix things like duplicates, spelling errors, etc.
```{r}
# Make DV factor
df$intensity <- as.factor(df$intensity)

# Fix other var types
df$year <- as.factor(df$year)
df$cname1 <- as.factor(df$cname1)
df$cabbrev1 <- as.factor(df$cabbrev1)
df$cname2 <- as.factor(df$cname2)
df$cabbrev2 <- as.factor(df$cabbrev2)

df$NATOmem_MEM <- factor(df$NATOmem_MEM, level = c("0", "1"), ordered = FALSE)
df$nato_stage <- factor(df$nato_stage, levels = c("Non-member", "PfP", "IntDial", "MAP", "Full"), ordered = FALSE)
df$nato_prememb <- factor(df$nato_prememb, level = c("0", "1"), ordered = FALSE)
df$nuclear1 <- factor(df$nuclear1, level = c("0", "1"), ordered = FALSE)
df$demo1 <- factor(df$demo1, level = c("0", "1"), ordered = FALSE)
df$demo2 <- factor(df$demo2, level = c("0", "1"), ordered = FALSE)
df$civilwar <- factor(df$civilwar, level = c("0", "1"), ordered = FALSE)

# Re-add continent info
df$continent <- countrycode::countrycode(df$ccode1, 'cown', 'continent')
df$continent[df$ccode1 == "372"] <- "Europe" # Georgia
df$continent[df$ccode1 == "347"] <- "Europe" # Kosovo
df$continent[df$ccode1 == "640"] <- "Europe" # Turkey
df$continent[df$ccode1 == "345"] <- "Europe" # Yugoslavia
df$continent[df$ccode1 == "315" | df$ccode1 == "316"] <- "Europe" # Czech Republic

# Drop microstates
df$gwcode <- countrycode::countrycode(df$ccode1, "cown", "gwn")

df <- df %>%
  dplyr::mutate(ccode1 = gwcode) %>%
  dplyr::select(-gwcode) %>%
  dplyr::filter(!is.na(ccode1))

# Log distance and GDP measures
# Add 1.01 to account for values of 0 in distance
df$lncapdist_us <- log(df$capdist_us + 1.01)
df$lnmindist_us <- log(df$mindist_us + 1.01)
df$lncentdist_us <- log(df$centdist_us + 1.01)

df$lncapdist_rus <- log(df$capdist_rus + 1.01)
df$lnmindist_rus <- log(df$mindist_rus + 1.01)
df$lncentdist_rus <- log(df$centdist_rus + 1.01)

df$lncapdist_natohq <- log(df$capdist_natohq + 1.01)
df$lncapdist_natorelative <- log(df$capdist_natorelative + 1.01 + abs(min(df$capdist_natorelative)))
df$lncapdist_natomin <- log(df$capdist_natomin + 1.01)

# Log gdp measures
df$lngdp1_2010const <- log(df$gdp1_2010const)
df$lngdp2_2010const <- log(df$gdp2_2010const)
df$lngdppc1_2010const <- log(df$gdppc1_2010const)
df$lngdppc2_2010const <- log(df$gdppc2_2010const)

# Log population measures
df$lnpop1 <- log(df$pop1 + 1.01)
df$lnpop2 <- log(df$pop2 + 1.01)

DT::datatable(plyr::count(df, "cname1"))
DT::datatable(plyr::count(df, "year"))

# Reorder df
df <- df %>%
  dplyr::select(year,
                cname1, cabbrev1, ccode1,
                cname2, cabbrev2, ccode2,
                continent, relevant_expansive, relevant_conserv,
                intensity,
                capdist_us, mindist_us, centdist_us, 
                lncapdist_us, lnmindist_us, lncentdist_us,
                capdist_rus, mindist_rus, centdist_rus,
                lncapdist_rus, lnmindist_rus, lncentdist_rus,
                capdist_natohq, capdist_natorelative,
                lncapdist_natohq, lncapdist_natorelative,
                capdist_natomin, lncapdist_natomin,
                NATOmem_MEM, NATOdur_MEM, nato_stage, nato_prememb,
                cinc1, cinc2, cinc_ratio,
                milex_sipri, milexpercap_sipri,
                nuclear1,
                gdp1_2010const, lngdp1_2010const, gdp2_2010const, lngdp2_2010const,
                gdppc1_2010const, lngdppc1_2010const, gdppc2_2010const, lngdppc2_2010const,
                pop1, lnpop1, pop2, lnpop2, 
                elf,
                age1,
                demo1, demo2, polity1, polity2, polity_intaxn,
                civilwar)

# Check missing
naniar::gg_miss_var(df)
visdat::vis_miss(df)

df %>%
  dplyr::filter(is.na(gdp1_2010const)) %>%
  DT::datatable()

df %>%
  dplyr::filter(is.na(milex_sipri | milexpercap_sipri)) %>%
  dplyr::select(year, cname1, milex_sipri, milexpercap_sipri) %>%
  DT::datatable()
```

# Save data
Save the final version of the data that is now prepped for modeling
```{r}
saveRDS(df, paste0(here::here(), "/data/","grayzone_model.rds"))
write.csv(df, paste0(here::here(), "/data/","grayzone_model.csv"))
```

# System info
This markdown file was last run and replicated on the following system
```{r}
Sys.info()
```
