==> docs/06b_Figures_Map.Rmd <==
---
title: "06b Figures: Map"
author: J Andres Gannon and Cole Reynolds
date: "`r Sys.Date()`"
output:
  html_document:
    theme: flatly
    code_download: yes
    code_folding: hide
    toc: yes
    number_sections: yes
    toc_float:
      collapsed: yes
  html_notebook:
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
editor_options:
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This document visualizes the geographic relationship between deterrence and the severity of Russian cyber operations.

# Preparation
## Load packages
Pipe operators have trouble loading for individual commands
```{r}
library(magrittr)
library(ggplot2)
```

## Load data
Load the cleaned and subset versions of each dataset
```{r}
# Load data
russia_cyber <- readRDS(paste0(here::here(), '/data/grayzone_model.rds'))
```

## Check data
```{r}
summary(russia_cyber)
```

# Prior data maps
Map the data using prior variable codings from other datasets

## DCID
Map of the DCID cases showing variation in severity
```{r}
dcid <- readRDS(file = paste0(here::here(),"/data/grayzone_dcid.rds"))

country <- unique(dcid$Target)
average_severity <- NULL 
for( i in 1:length(country)){
  average_severity[i] <- round(mean(dcid[dcid$Target == country[i], "Incident severity"]), 2)
}

dcid <- as.data.frame(cbind(country, average_severity), stringsAsFactors = F)
dcid$average_severity <- as.numeric(dcid$average_severity)

map <- rworldmap::joinCountryData2Map(dcid,
                                      nameJoinColumn="country",
                                      joinCode="NAME" )

rworldmap::mapDevice()

mapParams <- rworldmap::mapCountryData(map,
                                       nameColumnToPlot='average_severity',
                                       catMethod = c(2, 2.3, 2.6, 2.9, 3.2, 3.5),
                                       mapTitle = "\nRussian cyber attack severity (2005-2017) - Valeriano and Maness data",
                                       addLegend = FALSE,
                                       xlim = c(-167, 50),
                                       ylim = c(30,50)
                                      )
```

## REI
Map of the REI cases showing variation in severity. Info ops are coded as least severe, cyber disruption as moderate, and material support as most severe. Code each case by whatever is the most severe level that is present.
```{r}
rei <- readRDS(file = paste0(here::here(),"/data/grayzone_rei.rds"))

rei[rei$Target == "Ukraine ", "Target"] <- "Ukraine"

country <- unique(rei$Target)
average_severity <- NULL 
for( i in 1:length(country)){
  average_severity[i] <- round(mean(rei[rei$Target == country[i], "Favorable outcome"]), 2)
}

rei <- as.data.frame(cbind(country, average_severity), stringsAsFactors = F)
rei$average_severity <- as.numeric(rei$average_severity)

map <- rworldmap::joinCountryData2Map(rei,
                                      nameJoinColumn="country",
                                      joinCode="NAME" )

rworldmap::mapDevice()

mapParams <- rworldmap::mapCountryData(map,
                                       nameColumnToPlot = 'average_severity',
                                       catMethod = c(0, 0.4, 0.8, 1.2, 1.6, 2),
                                       mapTitle = "\nRussian cyber attack severity (1994-2017) - Way and Casey data",
                                       #colourPalette = c("lightgoldenrod", "yellow", "orange", "red"),
                                       addLegend = FALSE,
                                       xlim = c(-167, 50),
                                       ylim = c(30,50)
                                      )
```

## Combined
There should only be one case that is coded in both DCID and REI. For all the rest, scale their scores so we get a single score for each case in either dataset and we can compare across datasets. We know this is a rough coding since one is coding severity of the cyber attack and the other is coding the type of attack, but we'll start here.

# Current data maps
```{r}
rei_dcid <- readRDS(paste0(here::here(), "/data/grayzone_aggregate_cpass.rds"))

rei_dcid[rei_dcid$target == "Chechnya", "target"] <- "Russia"

rei_dcid[is.na(rei_dcid[,4]), 4] <- 0
rei_dcid[is.na(rei_dcid[,5]), 5] <- 0
rei_dcid[is.na(rei_dcid[,6]), 6] <- 0
rei_dcid[is.na(rei_dcid[,7]), 7] <- 0
rei_dcid[is.na(rei_dcid[,8]), 8] <- 0

rei_dcid$resp_convmil_gro <- rei_dcid$resp_convmil_gro * 5
rei_dcid$resp_convmil_airsea <- rei_dcid$resp_convmil_airsea * 4
rei_dcid$resp_paramil <- rei_dcid$resp_paramil * 3
rei_dcid$rei_cyberdisrup <- rei_dcid$rei_cyberdisrup * 2

for( i in 1:nrow(rei_dcid)){
  rei_dcid$intensity[i] <- max(rei_dcid$resp_convmil_gro[i], rei_dcid$resp_convmil_airsea[i], rei_dcid$resp_paramil[i], rei_dcid$resp_cyberdisrup[i], rei_dcid$resp_infoops[i])
}

country <- unique(rei_dcid$target)
average_severity <- NULL
for( i in 1:length(country)){
  average_severity[i] <- max(rei_dcid[rei_dcid$target == country[i], "intensity"])
  #average_severity[i] <- round(mean(dcid[dcid$Target == country[i], "severity_comparable"]), 2)
}

rei_dcid <- as.data.frame(cbind(country, average_severity), stringsAsFactors = F)
rei_dcid$average_severity <- as.numeric(rei_dcid$average_severity)
rei_dcid <- rbind(rei_dcid, c("Thailand", 2))

rei_dcid[rei_dcid$average_severity == 1, "average_severity"] <- "1 Info Ops"
rei_dcid[rei_dcid$average_severity == 2, "average_severity"] <- "2 Cyber Disruption"
rei_dcid[rei_dcid$average_severity == 3, "average_severity"] <- "3 Paramilitary"
rei_dcid[rei_dcid$average_severity == 4, "average_severity"] <- "4 Conventional Military (air/sea)"
rei_dcid[rei_dcid$average_severity == 5, "average_severity"] <- "5 Conventional Military (ground)"

map <- rworldmap::joinCountryData2Map(rei_dcid,
                                      nameJoinColumn="country",
                                      joinCode="NAME" )


mapParams <- rworldmap::mapCountryData(map,
                                       nameColumnToPlot='average_severity',
                                       catMethod = "categorical",
                                       #catMethod = c(0, 0.4, 0.8, 1.2, 1.6, 2),
                                       mapTitle = "\n\n\n\nIntensity of Russian activity (1994-2017)",
                                       addLegend = FALSE,
                                       xlim = c(-100,50), #c(-157, 50),
                                       ylim = c(40,70), #c(30,50),
                                       #colourPalette = RColorBrewer::brewer.pal(5, "Greys")
                                       colourPalette = c("grey85", "grey60", "grey40", "grey20", "grey0")
)

mapParams$legendText <- c("Info Ops", "Cyber Disruption", "Paramilitary", "Conventional Military (air/sea)", "Conventional Military (ground)")
```

# New map
Create new dummy df
```{r}
df <- russia_cyber %>%
  dplyr::filter(continent == "Europe") %>%
  dplyr::rename(target = cname1)
```

## Get geographic info
Only run once to get lat-lon data for each country to plot them
```{r, eval = FALSE}
library(ggmap)
mykey <- "AIzaSyAfKvl8754pRSsqRJgksUC9fOeq30bjda0"

register_google(key = mykey)

locations <- ggmap::mutate_geocode(df, target, key = mykey)

write.csv(locations, paste0(here::here(), "/data/","locations.csv"))
```

Merge new location info for each country
```{r}
locations <- read.csv(paste0(here::here(), "/data/","locations.csv")) %>%
  dplyr::select(year, target, lon, lat) %>%
  dplyr::mutate(year = as.factor(year))

df <- dplyr::left_join(df, locations)
```

## Wrangle
We want a single column that is the highest value for each country
```{r}
# Subset columns
df <- df %>%
  dplyr::select(target, intensity, lon, lat) %>%
  dplyr::rename(country = target)

df[is.na(df)] <- 0
df$intensity <- as.integer(df$intensity)

# Subset to the highest value for country
df <- df %>%
  dplyr::group_by(country) %>%
  dplyr::filter(intensity == max(intensity)) %>%
  dplyr::distinct() %>%
  dplyr::mutate(intensity = intensity - 1) 
```

## Create new map
We can now make a new map where each country is colored with the highest intensity value
```{r, fig.width = 7, fig.height = 4}
map <- df
map$intensity[map$intensity == 0] <- NA

# Create ISO codes, which have most reliable map shape files
map$iso <- countrycode::countrycode(map$country, "country.name", "iso3c")

# Add Kosovo and Yugoslavia (attack was in 1999, now Serbia and Montenegro) manually
map$iso[map$country == "Kosovo"] <- "RKS" # Does not plot, but irrelevant since intensity value = 0
map$iso[map$country == "Yugoslavia"] <- "SRB"

# Prep blank world map
world <- rnaturalearth::ne_countries(scale = "medium", 
                                     returnclass = "sf") %>%
  dplyr::select(iso_a3, geometry) %>%
  dplyr::rename(iso = iso_a3)

# Subset to European continent
world <- sf::st_crop(world, xmin = -20, xmax = 45,
                     ymin = 30, ymax = 73)

map <- dplyr::left_join(map, world) %>%
  dplyr::mutate(intensity = as.factor(intensity))

# Create color palette
palette <- RColorBrewer::brewer.pal(n = 5, 
                                    name = "Reds")

# Plot map
m <- ggplot(map) +
  geom_sf(aes(fill = intensity,
              geometry = geometry)) +
  scale_fill_manual(name = "", 
                    breaks = c(1, 2, 3, 4, 5), 
                    labels = c("Info Ops",
                               "Cyber Disruption",
                               "Paramilitary",
                               "Conventional Military",
                               "Conventional Military"),
                    values = palette) +
    labs(title = "European Targets of Russian Interventions (1994-2018)") +
    theme_minimal() +
    theme(title = element_text(size = 12),
          legend.text = element_text(size = 10),
          legend.position = "bottom")

m

ggsave("map_aggregate_europe.png", m, path = paste0(here::here(), '/paper/figures/'),
       width = 7.5, height = 5)
```
